refactor: move ResponseEntity creation to controllers

+++ src/main/java/com/yanapush/BrewerApp/service/UserServiceImpl.java
@@ -6,15 +6,11 @@ import com.yanapush.BrewerApp.entity.Recipe;
 import com.yanapush.BrewerApp.entity.User;
 import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
-import org.hibernate.HibernateException;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
+import org.springframework.security.authentication.BadCredentialsException;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
-
-import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -24,62 +20,48 @@ public class UserServiceImpl implements UserService, UserDetailsService {
     UserRepository repository;
 
     @Override
-    public ResponseEntity<?> getUser(int id) {
-        Optional<User> user = repository.findById(id);
-        return (user.equals(Optional.empty()))
-                ? new ResponseEntity<>(MessageConstants.ERROR_GETTING, HttpStatus.NOT_FOUND)
-                : new ResponseEntity<>(user, HttpStatus.OK);
+    public User getUser(int id) {
+        return repository.findById(id).orElseThrow(() -> new BadCredentialsException(MessageConstants.ERROR_GETTING));
     }
 
     @Override
     public User getUser(String username) {
-        Optional<User> user = repository.findByUsername(username);
-        return (user.equals(Optional.empty()))
-                ? null
-                : user.get();
+        return repository.findByUsername(username).orElseThrow(() -> new BadCredentialsException(MessageConstants.ERROR_GETTING));
     }
 
     @Override
-    public ResponseEntity<?> getUserByRecipe(Recipe recipe) {
-        Optional<User> user = repository.findByRecipesContains(recipe);
-        return (user.equals(Optional.empty()))
-                ? new ResponseEntity<>(MessageConstants.ERROR_GETTING, HttpStatus.NOT_FOUND)
-                : new ResponseEntity<>(user, HttpStatus.OK);
+    public User getUserByRecipe(Recipe recipe) {
+        return repository.findByRecipesContains(recipe).orElseThrow(() -> new BadCredentialsException(MessageConstants.ERROR_GETTING));
     }
 
     @Override
-    public ResponseEntity<?> addUser(User user) {
-        try {
-            repository.save(user);
-        } catch (HibernateException e) {
-            return new ResponseEntity<>(MessageConstants.ERROR_ADDING, HttpStatus.INTERNAL_SERVER_ERROR);
+    public boolean addUser(User user) {
+            return repository.save(user) == user;
+    }
+
+    @Override
+    public boolean deleteUser(User user) {
+        if (repository.existsById(user.getId())) {
+            repository.delete(user);
+            return !repository.existsById(user.getId());
         }
-        return new ResponseEntity<>(MessageConstants.SUCCESS_ADDING, HttpStatus.OK);
-    }
-
-    @Override
-    public void deleteUser(User user) {
-        repository.delete(user);
+        throw new BadCredentialsException(MessageConstants.ERROR_GETTING);
     }
 
     @Override
     public boolean deleteUser(int id) {
         if (repository.existsById(id)) {
             repository.deleteById(id);
-            return true;
+            return !repository.existsById(id);
         }
-        return false;
+        throw new BadCredentialsException(MessageConstants.ERROR_GETTING);
     }
 
     @Override
-    public ResponseEntity<?> changeUserPassword(String username, String password) {
-        Optional<User> user = repository.findByUsername(username);
-        if (user.equals(Optional.empty())) {
-            return new ResponseEntity<>(MessageConstants.ERROR_ADDING, HttpStatus.INTERNAL_SERVER_ERROR);
-        }
-        user.get().setPassword(password);
-        repository.save(user.get());
-        return new ResponseEntity<>(MessageConstants.SUCCESS_ADDING, HttpStatus.OK);
+    public boolean changeUserPassword(String username, String password) {
+        User user = repository.findByUsername(username).orElseThrow(() ->new BadCredentialsException(MessageConstants.ERROR_GETTING));
+        user.setPassword(password);
+        return repository.save(user) == user;
     }
 
     @Override
